Refactor to move in-memory uniqueness checker to test library
