+++ settings.gradle
@@ -355,7 +355,8 @@ include 'testing:p2p:inmemory-messaging-impl'
 include 'testing:sandboxes'
 include 'testing:security-manager-utilities'
 include 'testing:test-utilities'
-include 'testing:uniqueness-testkit'
+include 'testing:uniqueness:uniqueness-checker-fake'
+include 'testing:uniqueness:uniqueness-utilities'
 include 'testing:cpi-info-read-service-fake'
 include 'testing:virtual-node-info-read-service-fake'
 include 'tools:crypto-tck'
