+++ testing/uniqueness/uniqueness-utilities/build.gradle
@@ -2,7 +2,7 @@ plugins {
     id 'corda.common-library'
 }
 
-description 'Uniqueness checker testing library'
+description 'Uniqueness checker testing library utilities'
 
 dependencies {
     api "net.corda:corda-crypto"
